@@ -18,7 +18,7 @@
         }
     ],
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "ext-zip": "^1.14.0",
         "illuminate/console": "^10.10.0|^11.0",
         "illuminate/contracts": "^10.10.0|^11.0",
@@ -32,10 +32,10 @@
         "spatie/laravel-signal-aware-command": "^1.2|^2.0",
         "spatie/temporary-directory": "^2.0",
         "symfony/console": "^6.0|^7.0",
-        "symfony/finder": "^6.0|^7.0"
+        "symfony/finder": "^6.0|^7.0",
+        "ext-pcntl": "*"
     },
     "require-dev": {
-        "ext-pcntl": "*",
         "composer-runtime-api": "^2.0",
         "larastan/larastan": "^2.7.0",
         "laravel/slack-notification-channel": "^2.5|^3.0",
@@ -45,7 +45,8 @@
         "pestphp/pest": "^1.20|^2.0",
         "phpstan/extension-installer": "^1.1",
         "phpstan/phpstan-deprecation-rules": "^1.0",
-        "phpstan/phpstan-phpunit": "^1.1"
+        "phpstan/phpstan-phpunit": "^1.1",
+        "rector/rector": "^1.1"
     },
     "autoload": {
         "psr-4": {
@@ -61,9 +62,11 @@
         }
     },
     "scripts": {
-        "test": "vendor/bin/pest",
+        "test": "vendor/bin/pest --compact",
         "format": "vendor/bin/php-cs-fixer fix --allow-risky=yes",
-        "analyse": "vendor/bin/phpstan analyse"
+        "analyse": "vendor/bin/phpstan analyse",
+        "rector": "vendor/bin/rector --dry-run",
+        "baseline": "./vendor/bin/phpstan analyse --generate-baseline --memory-limit=2G"
     },
     "suggest": {
         "laravel/slack-notification-channel": "Required for sending notifications via Slack"
