@@ -8,6 +8,7 @@ use Generator;
 use Illuminate\Console\Command;
 use Illuminate\Support\Collection;
 use Spatie\Backup\BackupDestination\BackupDestination;
+use Spatie\Backup\Config\Config;
 use Spatie\Backup\Events\BackupManifestWasCreated;
 use Spatie\Backup\Events\BackupWasSuccessful;
 use Spatie\Backup\Events\BackupZipWasCreated;
@@ -27,8 +28,10 @@ class BackupJob
 
     protected FileSelection $fileSelection;
 
+    /** @var Collection<string, DbDumper> */
     protected Collection $dbDumpers;
 
+    /** @var Collection<int, BackupDestination> */
     protected Collection $backupDestinations;
 
     protected string $filename;
@@ -39,7 +42,7 @@ class BackupJob
 
     protected bool $signals = true;
 
-    public function __construct()
+    public function __construct(protected Config $config)
     {
         $this
             ->dontBackupFilesystem()
@@ -56,6 +59,7 @@ class BackupJob
         return $this;
     }
 
+    /** @param array<string> $allowedDbNames */
     public function onlyDbName(array $allowedDbNames): self
     {
         $this->dbDumpers = $this->dbDumpers->filter(
@@ -100,6 +104,10 @@ class BackupJob
         return $this;
     }
 
+    /**
+     * @param  Collection<string, DbDumper>  $dbDumpers
+     * @return $this
+     */
     public function setDbDumpers(Collection $dbDumpers): self
     {
         $this->dbDumpers = $dbDumpers;
@@ -127,6 +135,7 @@ class BackupJob
         return $this;
     }
 
+    /** @param Collection<int, BackupDestination> $backupDestinations */
     public function setBackupDestinations(Collection $backupDestinations): self
     {
         $this->backupDestinations = $backupDestinations;
@@ -134,12 +143,10 @@ class BackupJob
         return $this;
     }
 
-    /**
-     * @throws Exception
-     */
+    /** @throws Exception */
     public function run(): void
     {
-        $temporaryDirectoryPath = config('backup.backup.temporary_directory') ?? storage_path('app/backup-temp');
+        $temporaryDirectoryPath = $this->config->backup->temporaryDirectory ?? storage_path('app/backup-temp');
 
         $this->temporaryDirectory = (new TemporaryDirectory($temporaryDirectoryPath))
             ->name('temp')
@@ -170,7 +177,7 @@ class BackupJob
 
             $this->copyToBackupDestinations($zipFile);
         } catch (Exception $exception) {
-            consoleOutput()->error("Backup failed because: {$exception->getMessage()}." . PHP_EOL . $exception->getTraceAsString());
+            consoleOutput()->error("Backup failed because: {$exception->getMessage()}.".PHP_EOL.$exception->getTraceAsString());
 
             $this->temporaryDirectory->delete();
 
@@ -206,12 +213,13 @@ class BackupJob
         return $this->fileSelection->selectedFiles();
     }
 
+    /** @return array<string> */
     protected function directoriesUsedByBackupJob(): array
     {
         return $this->backupDestinations
             ->filter(fn (BackupDestination $backupDestination) => $backupDestination->filesystemType() === 'localfilesystemadapter')
             ->map(
-                fn (BackupDestination $backupDestination) => $backupDestination->disk()->path('') . $backupDestination->backupName()
+                fn (BackupDestination $backupDestination) => $backupDestination->disk()->path('').$backupDestination->backupName()
             )
             ->each(fn (string $backupDestinationDirectory) => $this->fileSelection->excludeFilesFrom($backupDestinationDirectory))
             ->push($this->temporaryDirectory->path())
@@ -222,7 +230,7 @@ class BackupJob
     {
         consoleOutput()->info("Zipping {$manifest->count()} files and directories...");
 
-        $pathToZip = $this->temporaryDirectory->path(config('backup.backup.destination.filename_prefix') . $this->filename);
+        $pathToZip = $this->temporaryDirectory->path($this->config->backup->destination->filenamePrefix.$this->filename);
 
         $zip = Zip::createForManifest($manifest, $pathToZip);
 
@@ -241,43 +249,44 @@ class BackupJob
      * Dumps the databases to the given directory.
      * Returns an array with paths to the dump files.
      *
-     * @return array
+     * @return array<string, string>
      */
     protected function dumpDatabases(): array
     {
         return $this->dbDumpers
-            ->map(function (DbDumper $dbDumper, $key) {
+            ->map(function (DbDumper $dbDumper, string $key): string {
                 consoleOutput()->info("Dumping database {$dbDumper->getDbName()}...");
 
-                $dbType = mb_strtolower(basename(str_replace('\\', '/', get_class($dbDumper))));
+                $dbType = mb_strtolower(basename(str_replace('\\', '/', $dbDumper::class)));
 
-
-                if (config('backup.backup.database_dump_filename_base') === 'connection') {
+                if ($this->config->backup->databaseDumpFilenameBase === 'connection') {
                     $dbName = $key;
                 } elseif ($dbDumper instanceof Sqlite) {
-                    $dbName = $key . '-database';
+                    $dbName = $key.'-database';
                 } else {
                     $dbName = $dbDumper->getDbName();
                 }
 
                 $timeStamp = '';
-                if ($timeStampFormat = config('backup.backup.database_dump_file_timestamp_format')) {
-                    $timeStamp = '-' . Carbon::now()->format($timeStampFormat);
+
+                if ($timeStampFormat = $this->config->backup->databaseDumpFileTimestampFormat) {
+                    $timeStamp = '-'.Carbon::now()->format($timeStampFormat);
                 }
 
                 $fileName = "{$dbType}-{$dbName}{$timeStamp}.{$this->getExtension($dbDumper)}";
 
+                // @todo is this still relevant or undocumented?
                 if (config('backup.backup.gzip_database_dump')) {
                     $dbDumper->useCompressor(new GzipCompressor());
-                    $fileName .= '.' . $dbDumper->getCompressorExtension();
+                    $fileName .= '.'.$dbDumper->getCompressorExtension();
                 }
 
-                if ($compressor = config('backup.backup.database_dump_compressor')) {
+                if ($compressor = $this->config->backup->databaseDumpCompressor) {
                     $dbDumper->useCompressor(new $compressor());
-                    $fileName .= '.' . $dbDumper->getCompressorExtension();
+                    $fileName .= '.'.$dbDumper->getCompressorExtension();
                 }
 
-                $temporaryFilePath = $this->temporaryDirectory->path('db-dumps' . DIRECTORY_SEPARATOR . $fileName);
+                $temporaryFilePath = $this->temporaryDirectory->path('db-dumps'.DIRECTORY_SEPARATOR.$fileName);
 
                 event(new DumpingDatabase($dbDumper));
 
@@ -315,7 +324,7 @@ class BackupJob
             });
     }
 
-    protected function sendNotification($notification): void
+    protected function sendNotification(object|string $notification): void
     {
         if ($this->sendNotifications) {
             rescue(
@@ -327,7 +336,7 @@ class BackupJob
 
     protected function getExtension(DbDumper $dbDumper): string
     {
-        if ($extension = config('backup.backup.database_dump_file_extension')) {
+        if ($extension = $this->config->backup->databaseDumpFileExtension) {
             return $extension;
         }
 
