@@ -9,27 +9,32 @@ use Symfony\Component\Finder\Finder;
 
 class FileSelection
 {
+    /** @var Collection<int, string> */
     protected Collection $includeFilesAndDirectories;
 
+    /** @var Collection<int, string> */
     protected Collection $excludeFilesAndDirectories;
 
     protected bool $shouldFollowLinks = false;
 
     protected bool $shouldIgnoreUnreadableDirs = false;
 
-    public static function create(array | string $includeFilesAndDirectories = []): self
+    /** @param array<string>|string $includeFilesAndDirectories */
+    public static function create(array|string $includeFilesAndDirectories = []): static
     {
         return new static($includeFilesAndDirectories);
     }
 
-    public function __construct(array | string $includeFilesAndDirectories = [])
+    /** @param array<string>|string $includeFilesAndDirectories */
+    public function __construct(array|string $includeFilesAndDirectories = [])
     {
         $this->includeFilesAndDirectories = collect($includeFilesAndDirectories);
 
         $this->excludeFilesAndDirectories = collect();
     }
 
-    public function excludeFilesFrom(array | string $excludeFilesAndDirectories): self
+    /** @param array<string>|string $excludeFilesAndDirectories */
+    public function excludeFilesFrom(array|string $excludeFilesAndDirectories): self
     {
         $this->excludeFilesAndDirectories = $this->excludeFilesAndDirectories->merge($this->sanitize($excludeFilesAndDirectories));
 
@@ -50,7 +55,8 @@ class FileSelection
         return $this;
     }
 
-    public function selectedFiles(): Generator | array
+    /** @return Generator|array<empty> */
+    public function selectedFiles(): Generator|array
     {
         if ($this->includeFilesAndDirectories->isEmpty()) {
             return [];
@@ -72,7 +78,7 @@ class FileSelection
             yield $includedFile;
         }
 
-        if (! count($this->includedDirectories())) {
+        if ($this->includedDirectories() === []) {
             return [];
         }
 
@@ -87,26 +93,29 @@ class FileSelection
         }
     }
 
+    /** @return array<string> */
     protected function includedFiles(): array
     {
         return $this
             ->includeFilesAndDirectories
-            ->filter(fn ($path) => is_file($path))->toArray();
+            ->filter(fn (string $path) => is_file($path))->toArray();
     }
 
+    /** @return array<string> */
     protected function includedDirectories(): array
     {
         return $this
             ->includeFilesAndDirectories
-            ->reject(fn ($path) => is_file($path))->toArray();
+            ->reject(fn (string $path) => is_file($path))->toArray();
     }
 
     protected function shouldExclude(string $path): bool
     {
         $path = realpath($path);
         if (is_dir($path)) {
-            $path .= DIRECTORY_SEPARATOR ;
+            $path .= DIRECTORY_SEPARATOR;
         }
+
         foreach ($this->excludeFilesAndDirectories as $excludedPath) {
             if (Str::startsWith($path, $excludedPath.(is_dir($excludedPath) ? DIRECTORY_SEPARATOR : ''))) {
                 if ($path != $excludedPath && is_file($excludedPath)) {
@@ -120,15 +129,19 @@ class FileSelection
         return false;
     }
 
-    protected function sanitize(string | array $paths): Collection
+    /**
+     * @param  string|array<string>  $paths
+     */
+    protected function sanitize(string|array $paths): Collection
     {
         return collect($paths)
-            ->reject(fn ($path) => $path === '')
-            ->flatMap(fn ($path) => $this->getMatchingPaths($path))
-            ->map(fn ($path) => realpath($path))
+            ->reject(fn (string $path) => $path === '')
+            ->flatMap(fn (string $path) => $this->getMatchingPaths($path))
+            ->map(fn (string $path) => realpath($path))
             ->reject(fn ($path) => $path === false);
     }
 
+    /** @return array<string> */
     protected function getMatchingPaths(string $path): array
     {
         if ($this->canUseGlobBrace($path)) {
@@ -140,6 +153,6 @@ class FileSelection
 
     protected function canUseGlobBrace(string $path): bool
     {
-        return strpos($path, '*') !== false && defined('GLOB_BRACE');
+        return str_contains($path, '*') && defined('GLOB_BRACE');
     }
 }
