@@ -4,6 +4,7 @@ namespace Spatie\Backup\Commands;
 
 use Exception;
 use Illuminate\Contracts\Console\Isolatable;
+use Spatie\Backup\Config\Config;
 use Spatie\Backup\Events\BackupHasFailed;
 use Spatie\Backup\Exceptions\BackupFailed;
 use Spatie\Backup\Exceptions\InvalidCommand;
@@ -18,6 +19,11 @@ class BackupCommand extends BaseCommand implements Isolatable
 
     protected $description = 'Run the backup.';
 
+    public function __construct(protected Config $config)
+    {
+        parent::__construct();
+    }
+
     public function handle(): int
     {
         consoleOutput()->comment($this->currentTry > 1 ? sprintf('Attempt n°%d...', $this->currentTry) : 'Starting backup...');
@@ -31,11 +37,12 @@ class BackupCommand extends BaseCommand implements Isolatable
         try {
             $this->guardAgainstInvalidOptions();
 
-            $backupJob = BackupJobFactory::createFromArray(config('backup'));
+            $backupJob = BackupJobFactory::createFromConfig($this->config);
 
             if ($this->option('only-db')) {
                 $backupJob->dontBackupFilesystem();
             }
+
             if ($this->option('db-name')) {
                 $backupJob->onlyDbName($this->option('db-name'));
             }
@@ -94,7 +101,7 @@ class BackupCommand extends BaseCommand implements Isolatable
         }
     }
 
-    protected function guardAgainstInvalidOptions()
+    protected function guardAgainstInvalidOptions(): void
     {
         if (! $this->option('only-db')) {
             return;
