@@ -5,6 +5,7 @@ namespace Spatie\Backup\Commands;
 use Exception;
 use Illuminate\Contracts\Console\Isolatable;
 use Spatie\Backup\BackupDestination\BackupDestinationFactory;
+use Spatie\Backup\Config\Config;
 use Spatie\Backup\Events\CleanupHasFailed;
 use Spatie\Backup\Tasks\Cleanup\CleanupJob;
 use Spatie\Backup\Tasks\Cleanup\CleanupStrategy;
@@ -20,13 +21,11 @@ class CleanupCommand extends BaseCommand implements Isolatable
     /** @var string */
     protected $description = 'Remove all backups older than specified number of days in config.';
 
-    protected CleanupStrategy $strategy;
-
-    public function __construct(CleanupStrategy $strategy)
-    {
+    public function __construct(
+        protected CleanupStrategy $strategy,
+        protected Config $config,
+    ) {
         parent::__construct();
-
-        $this->strategy = $strategy;
     }
 
     public function handle(): int
@@ -38,9 +37,7 @@ class CleanupCommand extends BaseCommand implements Isolatable
         $this->setTries('cleanup');
 
         try {
-            $config = config('backup');
-
-            $backupDestinations = BackupDestinationFactory::createFromArray($config['backup']);
+            $backupDestinations = BackupDestinationFactory::createFromArray($this->config);
 
             $cleanupJob = new CleanupJob($backupDestinations, $this->strategy, $disableNotifications);
 
