@@ -10,6 +10,5 @@ class BackupHasFailed
     public function __construct(
         public Exception $exception,
         public ?BackupDestination $backupDestination = null,
-    ) {
-    }
+    ) {}
 }
