@@ -6,6 +6,5 @@ class BackupZipWasCreated
 {
     public function __construct(
         public string $pathToZip,
-    ) {
-    }
+    ) {}
 }
