@@ -8,6 +8,5 @@ class CleanupWasSuccessful
 {
     public function __construct(
         public BackupDestination $backupDestination,
-    ) {
-    }
+    ) {}
 }
