@@ -8,6 +8,5 @@ class DumpingDatabase
 {
     public function __construct(
         public DbDumper $dbDumper
-    ) {
-    }
+    ) {}
 }
