@@ -8,6 +8,5 @@ class HealthyBackupWasFound
 {
     public function __construct(
         public BackupDestinationStatus $backupDestinationStatus,
-    ) {
-    }
+    ) {}
 }
