@@ -8,6 +8,5 @@ class UnhealthyBackupWasFound
 {
     public function __construct(
         public BackupDestinationStatus $backupDestinationStatus
-    ) {
-    }
+    ) {}
 }
