@@ -6,12 +6,14 @@ use Exception;
 
 class CannotCreateDbDumper extends Exception
 {
-    public static function unsupportedDriver(string $driver): self
+    public static function unsupportedDriver(string $driver): static
     {
-        $supportedDrivers = collect(config("database.connections"))->keys();
+        /** @var array<int, string> $supportedDrivers */
+        $supportedDrivers = config('database.connections');
 
-        $formattedSupportedDrivers = $supportedDrivers
-            ->map(fn (string $supportedDriver) => "`$supportedDriver`")
+        $formattedSupportedDrivers = collect($supportedDrivers)
+            ->keys()
+            ->map(fn (string $supportedDriver) => "`{$supportedDriver}`")
             ->join(glue: ', ', finalGlue: ' or ');
 
         return new static("Cannot create a dumper for db driver `{$driver}`. Use {$formattedSupportedDrivers}.");
