@@ -13,18 +13,17 @@ class BackupWasSuccessfulNotification extends BaseNotification
 {
     public function __construct(
         public BackupWasSuccessful $event,
-    ) {
-    }
+    ) {}
 
     public function toMail(): MailMessage
     {
         $mailMessage = (new MailMessage())
-            ->from(config('backup.notifications.mail.from.address', config('mail.from.address')), config('backup.notifications.mail.from.name', config('mail.from.name')))
+            ->from($this->config()->notifications->mail->from->address, $this->config()->notifications->mail->from->name)
             ->subject(trans('backup::notifications.backup_successful_subject', ['application_name' => $this->applicationName()]))
             ->line(trans('backup::notifications.backup_successful_body', ['application_name' => $this->applicationName(), 'disk_name' => $this->diskName()]));
 
         $this->backupDestinationProperties()->each(function ($value, $name) use ($mailMessage) {
-            $mailMessage->line("{$name}: $value");
+            $mailMessage->line("{$name}: {$value}");
         });
 
         return $mailMessage;
@@ -34,8 +33,8 @@ class BackupWasSuccessfulNotification extends BaseNotification
     {
         return (new SlackMessage())
             ->success()
-            ->from(config('backup.notifications.slack.username'), config('backup.notifications.slack.icon'))
-            ->to(config('backup.notifications.slack.channel'))
+            ->from($this->config()->notifications->slack->username, $this->config()->notifications->slack->icon)
+            ->to($this->config()->notifications->slack->channel)
             ->content(trans('backup::notifications.backup_successful_subject_title'))
             ->attachment(function (SlackAttachment $attachment) {
                 $attachment->fields($this->backupDestinationProperties()->toArray());
@@ -46,7 +45,7 @@ class BackupWasSuccessfulNotification extends BaseNotification
     {
         return (new DiscordMessage())
             ->success()
-            ->from(config('backup.notifications.discord.username'), config('backup.notifications.discord.avatar_url'))
+            ->from($this->config()->notifications->discord->username, $this->config()->notifications->discord->avatar_url)
             ->title(trans('backup::notifications.backup_successful_subject_title'))
             ->fields($this->backupDestinationProperties()->toArray());
     }
