@@ -8,18 +8,14 @@ use SplFileObject;
 
 class Manifest implements Countable
 {
-    protected string $manifestPath;
-
     public static function create(string $manifestPath): self
     {
         return new static($manifestPath);
     }
 
-    public function __construct(string $manifestPath)
+    public function __construct(protected string $manifestPath)
     {
-        $this->manifestPath = $manifestPath;
-
-        touch($manifestPath);
+        touch($this->manifestPath);
     }
 
     public function path(): string
@@ -27,7 +23,10 @@ class Manifest implements Countable
         return $this->manifestPath;
     }
 
-    public function addFiles(array | string | Generator $filePaths): self
+    /**
+     * @param  Generator|string|array<string>  $filePaths
+     */
+    public function addFiles(array|string|Generator $filePaths): self
     {
         if (is_string($filePaths)) {
             $filePaths = [$filePaths];
@@ -42,7 +41,7 @@ class Manifest implements Countable
         return $this;
     }
 
-    public function files(): Generator | array
+    public function files(): Generator|array
     {
         $file = new SplFileObject($this->path());
 
