@@ -9,8 +9,7 @@ class Period
     public function __construct(
         protected Carbon $startDate,
         protected Carbon $endDate
-    ) {
-    }
+    ) {}
 
     public function startDate(): Carbon
     {
