@@ -11,18 +11,18 @@ class BackupDestinationStatus
 {
     protected ?HealthCheckFailure $healthCheckFailure = null;
 
+    /** @param array<int, HealthCheck> $healthChecks */
     public function __construct(
         protected BackupDestination $backupDestination,
         protected array $healthChecks = []
-    ) {
-    }
+    ) {}
 
     public function backupDestination(): BackupDestination
     {
         return $this->backupDestination;
     }
 
-    public function check(HealthCheck $check): bool | HealthCheckFailure
+    public function check(HealthCheck $check): bool|HealthCheckFailure
     {
         try {
             $check->checkHealth($this->backupDestination());
@@ -33,6 +33,7 @@ class BackupDestinationStatus
         return true;
     }
 
+    /** @return Collection<int, HealthCheck> */
     public function getHealthChecks(): Collection
     {
         return collect($this->healthChecks)->prepend(new IsReachable());
