@@ -10,8 +10,7 @@ class MaximumAgeInDays extends HealthCheck
 {
     public function __construct(
         protected int $days = 1
-    ) {
-    }
+    ) {}
 
     public function checkHealth(BackupDestination $backupDestination): void
     {
