@@ -102,6 +102,7 @@ it('can backup using relative path', function () {
     foreach (range(0, $zip->numFiles - 1) as $i) {
         $zipFiles[] = $zip->statIndex($i)['name'];
     }
+
     $zip->close();
     sort($testFiles);
     sort($zipFiles);
@@ -121,9 +122,10 @@ it('can backup using short relative path', function () {
     if ($zip->numFiles) {
         $zipFile = $zip->statIndex(0)['name'];
     }
+
     $zip->close();
 
-    expect($zipFile)->toStartWith(ltrim($this->getStubDirectory(), DIRECTORY_SEPARATOR));
+    expect($zipFile)->toStartWith(ltrim((string) $this->getStubDirectory(), DIRECTORY_SEPARATOR));
 });
 
 it('excludes the temporary directory from the backup', function () {
@@ -132,6 +134,7 @@ it('excludes the temporary directory from the backup', function () {
     if (! file_exists($tempDirectoryPath)) {
         mkdir($tempDirectoryPath, 0777, true);
     }
+
     touch($tempDirectoryPath.DIRECTORY_SEPARATOR.'testing-file-temp.txt');
 
     $this->artisan('backup:run --only-files')->assertExitCode(0);
@@ -367,8 +370,10 @@ it('will encrypt backup when notifications are disabled', function () {
 
     $zip = new ZipArchive();
     $zip->open(Storage::disk('local')->path($this->expectedZipPath));
+
     expect($zip->numFiles)->toBe(1);
     expect($zip->statIndex(0)['encryption_method'])->toBe(ZipArchive::EM_AES_256);
+
     $zip->close();
 
     Event::assertNotDispatched(BackupZipWasCreated::class);
@@ -382,34 +387,42 @@ it('can use different compression methods for backup file', function () {
 
     $zip = new ZipArchive();
     $zip->open(Storage::disk('local')->path($this->expectedZipPath));
+
     expect($zip->numFiles)->toBe(1);
     expect($zip->statIndex(0)['comp_method'])->toBe(ZipArchive::CM_DEFLATE);
-    $zip->close();
 
+    $zip->close();
 
     // check no compression with ZipArchive::CM_STORE method
     config()->set('backup.backup.destination.compression_method', ZipArchive::CM_STORE);
     config()->set('backup.backup.destination.compression_level', 0);
 
+    \Spatie\Backup\Config\Config::rebind();
+
     $this->artisan('backup:run --only-db')->assertExitCode(0);
 
     $zip = new ZipArchive();
     $zip->open(Storage::disk('local')->path($this->expectedZipPath));
+
     expect($zip->numFiles)->toBe(1);
     expect($zip->statIndex(0)['comp_method'])->toBe(ZipArchive::CM_STORE);
-    $zip->close();
 
+    $zip->close();
 
     // check ZipArchive::CM_DEFLATE method with custom compression level
     config()->set('backup.backup.destination.compression_method', ZipArchive::CM_DEFLATE);
     config()->set('backup.backup.destination.compression_level', 2);
 
+    \Spatie\Backup\Config\Config::rebind();
+
     $this->artisan('backup:run --only-db')->assertExitCode(0);
 
     $zip = new ZipArchive();
     $zip->open(Storage::disk('local')->path($this->expectedZipPath));
+
     expect($zip->numFiles)->toBe(1);
     expect($zip->statIndex(0)['comp_method'])->toBe(ZipArchive::CM_DEFLATE);
+
     $zip->close();
 });
 
